pyproject.toml
[project]
name = "digit-recognition-system"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "numpy>=1.24.0",
    "pillow>=9.5.0",
    "streamlit>=1.25.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
]
